Change decimal to double

+++ YesPojiQuota.Core/Models/Quota.cs
@@ -4,7 +4,7 @@ namespace YesPojiQuota.Core.Models
     public class Quota
     {
         public int QuotaId { get; set; }
-        public decimal Available { get; set; }
+        public double Available { get; set; }
 
         public int AccountId { get; set; }
         public virtual Account Account { get; set; }
